revert change on move common file (2)
